[fixup] refactor vector reconciliation to get it correctly into the deployment for Stargate
