+++ pkg/stargate/deployments_test.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/k8ssandra/k8ssandra-operator/pkg/images"
 
+	testlogr "github.com/go-logr/logr/testing"
 	cassdcapi "github.com/k8ssandra/cass-operator/apis/cassandra/v1beta1"
 	api "github.com/k8ssandra/k8ssandra-operator/apis/stargate/v1alpha1"
 	"github.com/k8ssandra/k8ssandra-operator/pkg/encryption"
@@ -96,7 +97,8 @@ func TestNewDeployments(t *testing.T) {
 }
 
 func testNewDeploymentsDefaultRackSingleReplica(t *testing.T) {
-	deployments := NewDeployments(stargate, dc)
+	logger := testlogr.NewTestLogger(t)
+	deployments := NewDeployments(stargate, dc, logger)
 	require.Len(t, deployments, 1)
 	require.Contains(t, deployments, "cluster1-dc1-default-stargate-deployment")
 	deployment := deployments["cluster1-dc1-default-stargate-deployment"]
@@ -191,7 +193,8 @@ func testNewDeploymentsSingleRackManyReplicas(t *testing.T) {
 	stargate := stargate.DeepCopy()
 	stargate.Spec.Size = 3
 
-	deployments := NewDeployments(stargate, dc)
+	logger := testlogr.NewTestLogger(t)
+	deployments := NewDeployments(stargate, dc, logger)
 	require.Len(t, deployments, 1)
 	require.Contains(t, deployments, "cluster1-dc1-rack1-stargate-deployment")
 	deployment := deployments["cluster1-dc1-rack1-stargate-deployment"]
@@ -243,7 +246,8 @@ func testNewDeploymentsManyRacksManyReplicas(t *testing.T) {
 	stargate := stargate.DeepCopy()
 	stargate.Spec.Size = 8
 
-	deployments := NewDeployments(stargate, dc)
+	logger := testlogr.NewTestLogger(t)
+	deployments := NewDeployments(stargate, dc, logger)
 
 	require.Len(t, deployments, 3)
 	require.Contains(t, deployments, "cluster1-dc1-rack1-stargate-deployment")
@@ -339,7 +343,8 @@ func testNewDeploymentsManyRacksCustomAffinityDc(t *testing.T) {
 	stargate := stargate.DeepCopy()
 	stargate.Spec.Size = 8
 
-	deployments := NewDeployments(stargate, dc)
+	logger := testlogr.NewTestLogger(t)
+	deployments := NewDeployments(stargate, dc, logger)
 	require.Len(t, deployments, 3)
 	require.Contains(t, deployments, "cluster1-dc1-rack1-stargate-deployment")
 	require.Contains(t, deployments, "cluster1-dc1-rack2-stargate-deployment")
@@ -451,7 +456,8 @@ func testNewDeploymentsManyRacksCustomAffinityStargate(t *testing.T) {
 		},
 	}}
 
-	deployments := NewDeployments(stargate, dc)
+	logger := testlogr.NewTestLogger(t)
+	deployments := NewDeployments(stargate, dc, logger)
 	require.Len(t, deployments, 3)
 	require.Contains(t, deployments, "cluster1-dc1-rack1-stargate-deployment")
 	require.Contains(t, deployments, "cluster1-dc1-rack2-stargate-deployment")
@@ -526,7 +532,8 @@ func testNewDeploymentsManyRacksFewReplicas(t *testing.T) {
 	stargate := stargate.DeepCopy()
 	stargate.Spec.Size = 2 // rack3 will get no deployment
 
-	deployments := NewDeployments(stargate, dc)
+	logger := testlogr.NewTestLogger(t)
+	deployments := NewDeployments(stargate, dc, logger)
 	require.Len(t, deployments, 2)
 	require.Contains(t, deployments, "cluster1-dc1-rack1-stargate-deployment")
 	require.Contains(t, deployments, "cluster1-dc1-rack2-stargate-deployment")
@@ -547,7 +554,8 @@ func testNewDeploymentsCassandraConfigMap(t *testing.T) {
 	stargate := stargate.DeepCopy()
 	stargate.Spec.CassandraConfigMapRef = &corev1.LocalObjectReference{Name: configMapName}
 
-	deployments := NewDeployments(stargate, dc)
+	logger := testlogr.NewTestLogger(t)
+	deployments := NewDeployments(stargate, dc, logger)
 	require.Len(t, deployments, 1)
 	deployment := deployments["cluster1-dc1-default-stargate-deployment"]
 
@@ -579,7 +587,8 @@ func testNewDeploymentsEncryption(t *testing.T) {
 
 	stargate := stargate.DeepCopy()
 
-	deployments := NewDeployments(stargate, dc)
+	logger := testlogr.NewTestLogger(t)
+	deployments := NewDeployments(stargate, dc, logger)
 	require.Len(t, deployments, 1)
 	deployment := deployments["cluster1-dc1-default-stargate-deployment"]
 
@@ -683,7 +692,8 @@ func testNewDeploymentsAuthentication(t *testing.T) {
 	t.Run("disabled", func(t *testing.T) {
 		sg := stargate.DeepCopy()
 		sg.Spec.Auth = pointer.Bool(false)
-		deployments := NewDeployments(sg, dc)
+		logger := testlogr.NewTestLogger(t)
+		deployments := NewDeployments(sg, dc, logger)
 		require.Len(t, deployments, 1)
 		deployment := deployments["cluster1-dc1-default-stargate-deployment"]
 		javaOpts := utils.FindEnvVarInContainer(&deployment.Spec.Template.Spec.Containers[0], "JAVA_OPTS")
@@ -697,7 +707,8 @@ func testNewDeploymentsAuthentication(t *testing.T) {
 			ApiAuthMethod:   "Table",
 			TokenTtlSeconds: 123,
 		}
-		deployments := NewDeployments(sg, dc)
+		logger := testlogr.NewTestLogger(t)
+		deployments := NewDeployments(sg, dc, logger)
 		require.Len(t, deployments, 1)
 		deployment := deployments["cluster1-dc1-default-stargate-deployment"]
 		javaOpts := utils.FindEnvVarInContainer(&deployment.Spec.Template.Spec.Containers[0], "JAVA_OPTS")
@@ -712,7 +723,8 @@ func testNewDeploymentsAuthentication(t *testing.T) {
 			ApiAuthMethod:  "JWT",
 			JwtProviderUrl: "https://auth.example.com/auth/realms/stargate/protocol/openid-connect/token",
 		}
-		deployments := NewDeployments(sg, dc)
+		logger := testlogr.NewTestLogger(t)
+		deployments := NewDeployments(sg, dc, logger)
 		require.Len(t, deployments, 1)
 		deployment := deployments["cluster1-dc1-default-stargate-deployment"]
 		javaOpts := utils.FindEnvVarInContainer(&deployment.Spec.Template.Spec.Containers[0], "JAVA_OPTS")
@@ -728,7 +740,8 @@ func testImages(t *testing.T) {
 	t.Run("nil image 3", func(t *testing.T) {
 		stargate := stargate.DeepCopy()
 		stargate.Spec.ContainerImage = nil
-		deployments := NewDeployments(stargate, dc)
+		logger := testlogr.NewTestLogger(t)
+		deployments := NewDeployments(stargate, dc, logger)
 		require.Len(t, deployments, 1)
 		deployment := deployments["cluster1-dc1-default-stargate-deployment"]
 		assert.Equal(t, defaultImage3.String(), deployment.Spec.Template.Spec.Containers[0].Image)
@@ -740,7 +753,8 @@ func testImages(t *testing.T) {
 		stargate.Spec.ContainerImage = nil
 		dc := dc.DeepCopy()
 		dc.Spec.ServerVersion = "4.0.1"
-		deployments := NewDeployments(stargate, dc)
+		logger := testlogr.NewTestLogger(t)
+		deployments := NewDeployments(stargate, dc, logger)
 		require.Len(t, deployments, 1)
 		deployment := deployments["cluster1-dc1-default-stargate-deployment"]
 		assert.Equal(t, defaultImage4.String(), deployment.Spec.Template.Spec.Containers[0].Image)
@@ -753,7 +767,8 @@ func testImages(t *testing.T) {
 			Repository: "stargateio",
 			Tag:        "v" + DefaultVersion,
 		}
-		deployments := NewDeployments(stargate, dc)
+		logger := testlogr.NewTestLogger(t)
+		deployments := NewDeployments(stargate, dc, logger)
 		require.Len(t, deployments, 1)
 		deployment := deployments["cluster1-dc1-default-stargate-deployment"]
 		assert.Equal(t, defaultImage3.String(), deployment.Spec.Template.Spec.Containers[0].Image)
@@ -768,7 +783,8 @@ func testImages(t *testing.T) {
 		}
 		dc := dc.DeepCopy()
 		dc.Spec.ServerVersion = "4.0.1"
-		deployments := NewDeployments(stargate, dc)
+		logger := testlogr.NewTestLogger(t)
+		deployments := NewDeployments(stargate, dc, logger)
 		require.Len(t, deployments, 1)
 		deployment := deployments["cluster1-dc1-default-stargate-deployment"]
 		assert.Equal(t, defaultImage4.String(), deployment.Spec.Template.Spec.Containers[0].Image)
@@ -783,7 +799,8 @@ func testImages(t *testing.T) {
 			PullSecretRef: &corev1.LocalObjectReference{Name: "my-secret"},
 		}
 		stargate.Spec.ContainerImage = image
-		deployments := NewDeployments(stargate, dc)
+		logger := testlogr.NewTestLogger(t)
+		deployments := NewDeployments(stargate, dc, logger)
 		require.Len(t, deployments, 1)
 		deployment := deployments["cluster1-dc1-default-stargate-deployment"]
 		assert.Equal(t, "docker.io/my-custom-repo/stargate-3_11:latest", deployment.Spec.Template.Spec.Containers[0].Image)
@@ -801,7 +818,8 @@ func testImages(t *testing.T) {
 		stargate.Spec.ContainerImage = image
 		dc := dc.DeepCopy()
 		dc.Spec.ServerVersion = "4.0.1"
-		deployments := NewDeployments(stargate, dc)
+		logger := testlogr.NewTestLogger(t)
+		deployments := NewDeployments(stargate, dc, logger)
 		require.Len(t, deployments, 1)
 		deployment := deployments["cluster1-dc1-default-stargate-deployment"]
 		assert.Equal(t, "docker.io/my-custom-repo/stargate-4_0:latest", deployment.Spec.Template.Spec.Containers[0].Image)
