+++ pkg/telemetry/vector.go
@@ -10,24 +10,9 @@ import (
 	k8ssandraapi "github.com/k8ssandra/k8ssandra-operator/apis/k8ssandra/v1alpha1"
 	telemetry "github.com/k8ssandra/k8ssandra-operator/apis/telemetry/v1alpha1"
 	"github.com/k8ssandra/k8ssandra-operator/pkg/cassandra"
+	"github.com/k8ssandra/k8ssandra-operator/pkg/vector"
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-)
-
-// Default resources for Vector agent
-const (
-	DefaultVectorCpuRequest    = "100m"
-	DefaultVectorMemoryRequest = "128Mi"
-	DefaultVectorCpuLimit      = "2"
-	DefaultVectorMemoryLimit   = "2Gi"
-	DefaultScrapeInterval      = 30
-	// CassandraMetricsPortLegacy is the metrics port to scrape for the legacy MCAC stack (Metrics
-	// Collector for Apache Cassandra).
-	CassandraMetricsPortLegacy = 9103
-	// CassandraMetricsPortModern is the metrics port to scrape for the modern stack (metrics
-	// exposed by management-api).
-	CassandraMetricsPortModern = 9000
 )
 
 // InjectCassandraVectorAgent adds the Vector agent container to the Cassandra pods.
@@ -35,7 +20,7 @@ const (
 func InjectCassandraVectorAgent(telemetrySpec *telemetry.TelemetrySpec, dcConfig *cassandra.DatacenterConfig, k8cName string, logger logr.Logger) error {
 	if telemetrySpec.IsVectorEnabled() {
 		logger.Info("Injecting Vector agent into Cassandra pods")
-		vectorImage := DefaultVectorImage
+		vectorImage := vector.DefaultVectorImage
 		if telemetrySpec.Vector.Image != "" {
 			vectorImage = telemetrySpec.Vector.Image
 		}
@@ -53,7 +38,7 @@ func InjectCassandraVectorAgent(telemetrySpec *telemetry.TelemetrySpec, dcConfig
 			VolumeMounts: []corev1.VolumeMount{
 				{Name: "vector-config", MountPath: "/etc/vector"},
 			},
-			Resources: VectorContainerResources(telemetrySpec),
+			Resources: vector.VectorContainerResources(telemetrySpec),
 		}
 
 		logger.Info("Updating Vector agent in Cassandra pods")
@@ -77,29 +62,6 @@ func InjectCassandraVectorAgent(telemetrySpec *telemetry.TelemetrySpec, dcConfig
 	return nil
 }
 
-func VectorContainerResources(telemetrySpec *telemetry.TelemetrySpec) corev1.ResourceRequirements {
-	if telemetrySpec.Vector.Resources != nil {
-		return *telemetrySpec.Vector.Resources
-	}
-
-	return corev1.ResourceRequirements{
-		Requests: corev1.ResourceList{
-			corev1.ResourceCPU:    resource.MustParse(DefaultVectorCpuRequest),
-			corev1.ResourceMemory: resource.MustParse(DefaultVectorMemoryRequest),
-		},
-		Limits: corev1.ResourceList{
-			corev1.ResourceMemory: resource.MustParse(DefaultVectorMemoryLimit),
-			corev1.ResourceCPU:    resource.MustParse(DefaultVectorCpuLimit),
-		},
-	}
-}
-
-type VectorConfig struct {
-	Sinks          string
-	ScrapePort     int32
-	ScrapeInterval int32
-}
-
 func CreateCassandraVectorToml(telemetrySpec *telemetry.TelemetrySpec, mcacEnabled bool) (string, error) {
 	vectorConfigToml := `
 [sinks.console]
@@ -117,17 +79,17 @@ target = "stdout"
 
 	var scrapePort int32
 	if mcacEnabled {
-		scrapePort = CassandraMetricsPortLegacy
+		scrapePort = vector.CassandraMetricsPortLegacy
 	} else {
-		scrapePort = CassandraMetricsPortModern
+		scrapePort = vector.CassandraMetricsPortModern
 	}
 
-	var scrapeInterval int32 = DefaultScrapeInterval
+	var scrapeInterval int32 = vector.DefaultScrapeInterval
 	if telemetrySpec.Vector.ScrapeInterval != nil {
 		scrapeInterval = int32(telemetrySpec.Vector.ScrapeInterval.Seconds())
 	}
 
-	config := VectorConfig{
+	config := vector.VectorConfig{
 		Sinks:          vectorConfigToml,
 		ScrapePort:     scrapePort,
 		ScrapeInterval: scrapeInterval,
